requirements.txt
fastapi
uvicorn
jinja2
python-multipart
httpx
google-generativeai
python-dotenv
